Support c10::optional<Tensor> in custom C++ autograd function.
Certain autograd functions can have optional Tensor arguments. For
this purpose it would be nice to support c10::optional<Tensor> as an argument
for C++ autograd functions.

I've added the appropriate overload to ExtractVariables to ensure this works.
For an example, you can look at D21272807 in terms of how this is used.

Differential Revision: [D21363491](https://our.internmc.facebook.com/intern/diff/D21363491/)

ghstack-source-id: 103343057
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37700